ep5-code restructure
