+++ index.html
@@ -11,6 +11,6 @@
         <div id="root">
             <h1>NOT RENDER!</h1>
         </div>
-        <script type="module" src="./App.jsx"></script>
+        <script type="module" src="./src/App.jsx"></script>
     </body>
 </html>
