+++ notes/ep5.md
@@ -1,0 +1,17 @@
+# code modularity
+Modularity is the practice of dividing your code into smaller, independent, and cohesive units that perform specific tasks or functions. These units are often called modules, components, or classes, depending on the programming language and paradigm you use. Modularity allows you to organize your code logically, avoid repetition, and facilitate testing and debugging.
+- divide code into smaller component, one component should not exceed the 100-200 lines of code.
+- put all you hardcoded data into seperate file.
+
+# folder structure
+Always try to follow industry standard folder structure
+- e.g: https://legacy.reactjs.org/docs/faq-structure.html
+
+# export
+- default export
+    - export default **functionname**
+- named export (when we need to export multiple functions/objects from file)
+    - export **function/object name** / export {objects, functions}
+
+
+
