+++ src/Components/Search.jsx
@@ -1,0 +1,7 @@
+const Search = () => {
+    return (
+        <div className="search-container">Search</div>
+    )
+}
+
+export default Search
