+++ src/Utils/constant.js
@@ -1,0 +1,2 @@
+export const CDN_URL = "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/"
+export const LOGO_URL = "https://seeklogo.com/images/F/food-logo-59E5A73AFD-seeklogo.com.png"
